LayoutTests/fullscreen/viewer-mode-should-block-element-fullscreen-expected.txt


EVENT(canplay)
RUN(window.internals.enterViewerMode(video))
EVENT(fullscreenchange)
RUN(document.body.requestFullscreen())
Promise rejected correctly OK
END OF TEST
